requirements.txt
beautifulsoup4
feedparser
jinja2
lxml
pangu
playwright
Pillow
python-dotenv
requests
tqdm
youtube-transcript-api
Flask